fix(snackbar): differentiated clear highlight
Addresses issue #30. The snackbar usually just displays messages, so for
calling the user for an action like 'clear highlighted words', the
snackbar should appear distinct, which it does now.

There is a minor issue if you are on a page with highlighted words and
you then click on a link to the headline, the 'copied link' message will
also appear in the same style. The right thing to do would be to
implement stacking notifications, or place the 'clear highlights' action
somewhere else entirely.