fix: regression for TOC styling
Unfortunately, it's not possible to automatically add a label "Contents"
for the toctree with CSS alone.

When the :hidden: option is not used, it works as desired, but when
:hidden: is used, the .toctree-wrapper div is not truly empty (it
contains a linebreak).

That means, I can't hide the caption with the :empty pseudo-selector,
because in current browser implementations, :empty does not include
whitespace.

I'll have to revert to explitly including a heading (or better, a
rubric) before using toctree directives without :hidden:. I may enable
that again later, if I find a way to 'post-process' the stuff written
out by Sphinx.