fix: tests for headerlinks mechanism (#337)
* fix: tests for headerlinks mechanism

Sphinx 3.5 replaced the `html_add_permalinks` configuration option
with `html_permalinks` and `html_permalinks_icon`. This commit switches
over to the new mechanism, including fixing the broken tests,
but it will likely crash with Sphinx versions <3.5. Fixes issue #336.

* fix: update netlify deps

* fix: parametrize sphinx version for tests and docs

Building docs fails with Sphinx <3.5 due to the different headerlinks
configuration. Tests, that were changed with the last commit, now also
fail with Sphinx <3.5.

* fix: backport for permalinks

I make the `html_permalinks` option from Sphinx 3.5 available
in this theme. If the Sphinx version is older than 3.5, this theme
exposes the new option, so that this version of the theme will
continue to work with older versions of Sphinx.