chore: mypy update (#338)
Mypy 0.8 changed some behavior in finding the source files.
Previously, I added the whole `tests` directory for checking
in the noxfile. This caused problems because of duplicate `conf.py`
files in the tests' 'root' directories.

Now, I pass a list of files in the `mypy.ini` section, being a bit
more explicit.