Feat: Use CSS Grid for layout (#307)
* Feat: Use CSS Grid for layout

First step of a series of structural improvements. Using CSS grid enables
to get rid of a few wrapper `div` elements and make the structure
overall simpler.

* feat: minor tweaks to fix new layout

* feat: restructuring

* feat: further tweaks in restructuring

Now, the sidebar opens and closes instead of the nav within.
Users should just be able to overwrite the sidebar block while
still keeping the functionality.

* feat: use tailwind 2 colors

* fix: sidebar issues

* fix: snackbar

Turned the snackbar into a button. The only real use case is to
clear the highlighted search terms. Renamed the 'nav' template
to 'sidebar'.

* fix: minor tweaks

* fix: minor tweaks

* fix: minor tweaks

* fix: pyproject toml build update

* fix: restructuring tweak

* fix: clear highlights

* feat: scroll to top button

* Feat: Use InteractionObserver API

to make the currently visible section `active`.
This is supposedly the right way of doing things instead
of firing a lot of scroll events.

* feat: scroll to top improvements

* fix: scroll to top focus and hover styles

* fix: make scroll to top optional

* fix: intersection observer tweak

I adjusted `rootMargin` on the intersection observer to only have
one active class at a time. This still doesn't work quite as I'd like.

* fix: remove word 'Copyright'

It's duplicate information to print the copyright symbol and the word
'Copyright'. Let's remove it.

* fix: tweaks to sidebar width and margin

I adjusted the sidebar width and the margin of the main element.

* fix: test

* fix: rounded none for inputs

* fix: remove unused Tailwind animations

* fix: intersection observer tweaks

* fix: intersection observer tweaks