docs: update (#491)
* fix: mypy errors

I had to build in checks, if `app.builder is not None` to silence
mypy. Not too happy about this, but the logic is, that the code
might error out, because `app.builder` is apparently of type `Optional`
and if it's `None`, then no methods are defined on it.

* docs: update installation instructions

Changed the file format to `.md` and edited the installation
instructions a little. The main part is adding the installation
instructions for Poetry, so that people don't have to go to different
pages just for installing the dependencies.

* fix: footnote styles

- removed margin-left on footnote number
- removed duplicate horizontal line on footnote list