Style tweaks (#523)
* feat: tweak page title style

- reduced font size and make it a degree bolder

* fix: remove border on sidebar

I think there isn't a need for a border because of the ample
amount of whitespace between the sidebar and the main text.

* fix: navigation link style

I made the links smaller, bolder, and less colorful.

* fix: tweak admonitions

Add a 1px border radius and a subtle background color matching
their border.

* fix: border radius to code blocks

It seems like I'm going back and forth between having a border radius or
not. The effect is very subtle but a 1px border radius does take the
edge off a bit.

I also added a background color to code blocks.

* fix: inline code

* fix: inline code

* fix: shadows and border radius

- added box shadow to admonitions and code blocks
- increased their border radius to 2px

* fix: vale action

* fix: removed borders of admonitions

Because of using a box shadow, the borders weren't necessary any more.

* fix: tweak code styles

* fix: inline code styles

* fix: better font handling (#524)

The fonts are in an npm package instead of manually copied.
This makes changing and updating the fonts a lot easier.
I also swapped `Roboto Mono` with `JetBrains Mono` which I think
looks fabulous.

* fix: decrease font size for inline code

* fix: tweak tooltips

* fix: tests

* feat: kbd style

* fix: header layout

I tweaked the header layout a bit to prepare for additional site
links on the right of the header.

* fix: no copy button in Python code

I want to add them later in JavaScript to the `<pre>` tag directly.
The copy functionality requires JavaScript anyway. Should there be
any issue with the JavaScript, I'd rather the button doesn't render
at all.

* fix: link color

I created two variables `color-highlight` and `color-link` which
are used consistently. This allows a user to override these easily
with custom CSS. I also changed the `highlight` color to some light
blue, as I prefer this color. I don't remember why I used pink in the
first place. I don't even like pink.

* feat: add logo

The icon's name is 'auto awesome'. After finding out that material
icons has something like this, I just had to use it.

* fix: broken tests

Some of the tests are broken but I don't care to fix this now.
I want to rethink the testing strategy for both the theme and
the docs website from first principles.

* feat: move to Stimulus (#522)

This PR adds the Stimulus framework for better JavaScript organization.

* fix: change color name

from `color-highlight` to `color-brand`

* fix: input style tweaks

* feat: copy code block to clipboard

The 'copy code' function is back! This time in JavaScript.
The rationale is that the button shouldn't even render if there's
any issue with the JavaScript code. I also got rid of the tooltip and
icon. A clear label is superior.

* feat: add favicons

* fix: remove ids

Since the move to Stimulus, I don't have to use this many IDs
since the JavaScript is now connected to data-attributes which I
can also use for styling. This way, the change in styling is also
connected to the element with the data-attribute.

* fix: search input tweak

* docs: font addition

* fix: search input tweak

* fix: scroll to top styles

* feat: better template layouts

* feat: disable navigation with keys

* fix: copy button appears on focus

* feat: redesign  (#531)

* fix: header layout
* feat: redesign

* fix: minor tweaks

* fix: spacing

* fix: toctree styles

* fix: figures and images style

* fix: table styles

* fix: add image source

* fix: remove empty toctree-wrappers

* fix: autodocs

* fix: highlighted inline code

* fix: Python versions

* fix: code yellow

* fix: copy button sticks

* feat: simple placeholder highlighting

I can get rid of the special `samp` directive now.
I implemented a simple regex based version that
works properly with code blocks. No need for a special
directive anymore.

I also removed the special treatment of `console` and
will use `shell` from now on everywhere.

* fix: eliminate traces of samp directive usage

Removed the sphinx samp directive from dependencies.

* fix: add changelog

* fix: placeholder tags

Google recommends `<code><var>` tags for placeholder text.
That's what I'm going to use from now on.

* docs: update install

* docs: minor edit

* fix: headerlinks

* chore: package update