fix: canonical links (#1262)
Upstream Sphinx doesn't construct canonical links correctly for the
DirectoryHTMLBuilder.
When I introduced #650, it included a bug for `index.html` pages.

This PR adds a proper fix with tests this time to make sure that
canonical links are what they should be. This PR is a follow up of #1258
and reverts the change in `layout.html`. Checking the canonical link
shouldn't be done in the template.

It's reported here:
[#9730](https://github.com/sphinx-doc/sphinx/issues/9730)