feat: extrahead  (#1267)
Add a new block to the `layout.html` template just before the closing
`</head>`.
`{% block extrahead %}{% end %}`. 

This can be used to add any additional tags, for example analytics
`<script>` tags.

The downside is that users would have to copy the whole `page.html`
template to add the block.
Maybe I'll find a better way in the future.