fix: re-implementing emhasizing placeholders in code blocks
Implement the core piece as a Pygments filter.
Then, override the `PygmentsBridge.highlight_block` method to add the
filter to the lexer.
This is comparatively simpler than overriding the `visit_literal_block` method
of the HTMLTranslator in Sphinx.
This commit also adds the code buttons back to the code blocks.