feat: restyle everything (#1292)
Where to even begin? 

This PR is big but several ideas coincided that lead to this.

## Approach

I knew that back then, but copy/pasting whole methods from
Sphinx/Docutils to override them isn't worth their effect. As much as I
can, I just use the defaults and accept that the DOM might not be
_exactly_ what I want. That's why I originally used the `postprocessing`
approach to modify only the HTML DOM _after_ rendering.

## Python

I trimmed down the custom `HTMLTranslator` to just one method
`depart_reference` which I use to add an external link icon after
external reference links. I added an option for this, which is _False_
by default.

I implemented the `emphasize-text` option as a Pygments filter. Although
I have to override the `highlight_block` method for this, it's much
simpler than the `visit_literal_block` method.

The custom `HTMLFormatter` is now much simpler, just overriding the
`_highlight_lines` method. I removed all the custom wrapping and instead
use the `wrapcode` option of the Pygments HTML formatter.

The Post-processing code grew a little, but it's much simpler to add the
programming language to the code block headers in the HTML than it is to
modify the doctree in various places.

## Styles

I added the Tailwind Typography plugin and re-implemented the styles
using the Tailwind configuration. Although the CSS size is the same
(probably due to the lengthy selectors from Tailwind's Typography
plugin), I feel like there are much fewer and cleaner classes to achieve
the same look and feel.
All SVG icons have now `height` and `width` attributes to size them,
even when the CSS isn't immediately available.