feat: simplify highlighting (#1329)
Instead of copying and pasting the `highlight_block` method from Sphinx,
I just monkey-patch it to add the relevant options,
which results in less code.

This PR also gets rid of the `icons.py` file, because it was only used
in the `postprocess` methods. Instead of a dictionary, we store the 2
icons in a dataclass.