fix: pin Sphinx version (#1511)
In Sphinx 7, copying the static assets happens BEFORE the
`doctree-resolved` event is emitted.
Because we update the globalcontext with the DocSearch configuration
when handling this event, the `docsearch_config.js` file ends up empty.
DocSearch is NOT being added.

In Sphinx 6, copying the static files happens AFTER the
`doctree-resolved` event is emitted.
The global context has the DocSearch configuration, the
`docsearch_config.js` file is NOT empty, and DocSearch works as
expected.

Properly supporting the latest version of Sphinx is a bit of a challenge
due to the many expected and unexpected changes. Until I figure it out,
I won't support Sphinx 7.