fix: style tweak for auto api (#2057)
Fix display issues for type hints in API documentation with
`sphinx-autoapi`.
To be able to better check this, this PR also adds a small mock Python
module,
with a demo page for documenting it. 

Styling is not 100% perfect, as the page now includes double lines (one
for the heading 'Module contents' and one for the function signature).
But still better than before.

@vict0rsch I hope it's ok that I took your function from issue #2056 and
put it as an example in these docs.

Fixes #2056.